Refactor import error handling in script_helpers.sh to ensure state object removal only occurs if import results are available

+++ deploy/scripts/helpers/script_helpers.sh
@@ -580,14 +580,16 @@ function ImportAndReRunApply {
 					# shellcheck disable=SC2086
 					if ! terraform -chdir="${terraform_module_directory}" import $importParameters -no-color "${moduleID}" "${azureResourceID}" | tee -a import_result.txt; then
 						import_return_value=${PIPESTATUS[0]}
-						moduleName=$(grep "^module..*.." .import_result.txt | cut -d' ' -f1 | sed 's/.$//')
-						echo "Removing state object:           ${moduleName}"
-						if terraform -chdir="${terraform_module_directory}" state rm "${moduleName}"; then
+						if [ -f import_result.txt ]; then
+							moduleName=$(grep "^module..*.." import_result.txt | cut -d' ' -f1 | sed 's/.$//')
+							echo "Removing state object:           ${moduleName}"
+							if terraform -chdir="${terraform_module_directory}" state rm "${moduleName}"; then
 
-							if ! terraform -chdir="${terraform_module_directory}" import $importParameters -no-color "${moduleID}" "${azureResourceID}"; then
-								import_return_value=$?
-							else
-								import_return_value=0
+								if ! terraform -chdir="${terraform_module_directory}" import $importParameters -no-color "${moduleID}" "${azureResourceID}"; then
+									import_return_value=$?
+								else
+									import_return_value=0
+								fi
 							fi
 						fi
 					fi
